Properly fixed layout bug that was "fixed" in the last commit
I figured out what was causing the bug and fixed it accordingly. the
ul element comes with a default padding which was messing the layout of
the li elements, since I'm having them in an inline block position.
Setting the padding to 0 allowed me to correctly set the layout.
Had to learn to use :last child feature to prevent the rightmost element
from messing the layout with a margin to the right.